(BKR-1560) Add install_puppet_agent_from_dev_builds_on
In puppet-agent's test suite, agent packages are initially installed via
a reusable setup test at setup/aio/010_Install_Puppet_Agent.rb, which
installs an internal development build of puppet-agent by git sha or
tag. This change adds a new helper method which mimics that behavior.

The `install_puppet_agent_on` helper is updated to use this method when
the dev builds are accessible, so that it can install puppet-agent on
platforms that are only supported within PE.